+++ TowerDefence/TowerDefence/Vector2D.h
@@ -3,10 +3,10 @@
 
 struct Vector2D
 {
-	int x;
-	int y;
+	float x;
+	float y;
 
-	Vector2D(int x, int y) : x(x), y(y) {}
+	Vector2D(float x, float y) : x(x), y(y) {}
 
 	static Vector2D Zero() { return Vector2D(0, 0); }
 	static Vector2D One() { return Vector2D(1, 1); }
